+++ spec/unit/braintree/webhook_notification_gateway_spec.js
@@ -437,6 +437,22 @@ describe('WebhookNotificationGateway', function () {
       });
     });
 
+    it('builds a sample notification for an OAuth access revocation webhook', function (done) {
+      let notification = specHelper.defaultGateway.webhookTesting.sampleNotification(
+        WebhookNotification.Kind.OAuthAccessRevoked,
+        'my_id'
+      );
+      let bt_signature = notification.bt_signature;
+      let bt_payload = notification.bt_payload;
+
+      specHelper.defaultGateway.webhookNotification.parse(bt_signature, bt_payload, function (err, webhookNotification) {
+        assert.equal(webhookNotification.kind, WebhookNotification.Kind.OAuthAccessRevoked);
+        assert.equal(webhookNotification.oauthAccessRevocation.merchantId, 'abc123');
+        assert.exists(webhookNotification.timestamp);
+        done();
+      });
+    });
+
     it('builds a sample notification for a connected merchant status transitioned webhook', (done) => {
       let notification = specHelper.defaultGateway.webhookTesting.sampleNotification(
         WebhookNotification.Kind.ConnectedMerchantStatusTransitioned,