+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Add `oauthAccessRevocation` to `WebhookNotification`s
+
 ## 2.8.0
 * Fix binary multipart uploads.
 * Add support for `association_filter_id` in `Customer#find`